requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
caugetch==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
clipboard==0.0.4
colorama==0.4.6
elastic-transport==8.4.0
elasticsearch==8.5.2
frozenlist==1.3.3
getpass4==0.0.14.1
idna==3.4
multidict==6.0.2
pyaes==1.6.1
pyasn1==0.4.8
pyperclip==1.8.2
python-dateutil==2.8.2
requests==2.28.1
rsa==4.9
six==1.16.0
Telethon==1.26.0
urllib3==1.26.13
yarl==1.8.1
